+++ shepherd_core/tests/test_writer.py
@@ -5,11 +5,13 @@ from typing import Union
 import numpy as np
 import pytest
 
+from shepherd_core import CalibrationPair
 from shepherd_core import Reader
 from shepherd_core import Writer
 from shepherd_core.data_models import CalibrationEmulator as CalEmu
 from shepherd_core.data_models import CalibrationHarvester as CalHrv
 from shepherd_core.data_models import CalibrationSeries as CalSeries
+from shepherd_core.data_models import EnergyDType
 from shepherd_core.data_models.task import Compression
 
 
@@ -90,22 +92,35 @@ def test_writer_faulty_path(tmp_path: Path) -> None:
         generate_shp_file(tmp_path)
 
 
-def test_writer_modify(h5_file: Path) -> None:
-    with Writer(h5_file, modify_existing=True) as sfw:
-        sfw.store_hostname("sheep")
-
-
 def test_writer_init_cal_emu(h5_path: Path) -> None:
-    generate_shp_file(h5_path, cal_data=CalEmu())
+    vgain = 0.123
+    generate_shp_file(h5_path, cal_data=CalEmu(dac_V_A=CalibrationPair(gain=vgain)))
+    # dac_V_A translates to .ds_voltage, should stay untouched
+    with Reader(h5_path) as sfr:
+        assert sfr.ds_voltage.attrs["gain"] == vgain
+    with Writer(h5_path, modify_existing=True) as sfw:
+        assert sfw.ds_voltage.attrs["gain"] == vgain
 
 
 def test_writer_init_cal_hrv(h5_path: Path) -> None:
-    generate_shp_file(h5_path, cal_data=CalHrv())
+    vgain = 0.123
+    generate_shp_file(h5_path, cal_data=CalHrv(adc_V_Sense=CalibrationPair(gain=vgain)))
+    # adc_V_Sense translates to .ds_voltage, should stay untouched
+    with Reader(h5_path) as sfr:
+        assert sfr.ds_voltage.attrs["gain"] == vgain
+    with Writer(h5_path, modify_existing=True) as sfw:
+        assert sfw.ds_voltage.attrs["gain"] == vgain
 
 
 def test_writer_init_cal_series(h5_path: Path) -> None:
-    cs = CalSeries.from_cal(CalHrv())
+    vgain = 0.123
+    cs = CalSeries.from_cal(CalHrv(adc_V_Sense=CalibrationPair(gain=vgain)))
     generate_shp_file(h5_path, cal_data=cs)
+    # adc_V_Sense translates to .ds_voltage, should stay untouched
+    with Reader(h5_path) as sfr:
+        assert sfr.ds_voltage.attrs["gain"] == vgain
+    with Writer(h5_path, modify_existing=True) as sfw:
+        assert sfw.ds_voltage.attrs["gain"] == vgain
 
 
 def test_writer_faulty_window(h5_path: Path) -> None:
@@ -161,6 +176,114 @@ def test_writer_setter(h5_path: Path) -> None:
     with Reader(h5_path) as sfr:
         assert sfr.get_hostname() == name
 
+
+def test_writer_modify_mode(h5_path: Path) -> None:
+    generate_shp_file(h5_path, mode="harvester")
+    with Reader(h5_path) as sfr:
+        assert sfr.get_mode() == "harvester"
+    with Writer(h5_path, modify_existing=True, mode="emulator") as sfw:
+        assert sfw.get_mode() == "emulator"
+    with Reader(h5_path) as sfr:
+        assert sfr.get_mode() == "emulator"
+
+
+def test_writer_not_modify_mode(h5_path: Path) -> None:
+    generate_shp_file(h5_path, mode="harvester")
+    with Reader(h5_path) as sfr:
+        assert sfr.get_mode() == "harvester"
+    with Writer(h5_path, modify_existing=True) as sfw:
+        assert sfw.get_mode() == "harvester"
+    with Reader(h5_path) as sfr:
+        assert sfr.get_mode() == "harvester"
+
+
+def test_writer_modify_datatype(h5_path: Path) -> None:
+    generate_shp_file(h5_path, datatype="ivcurve", window_samples=1024)
+    with Reader(h5_path) as sfr:
+        assert sfr.get_datatype() == EnergyDType.ivcurve
+    with Writer(h5_path, modify_existing=True, datatype="ivsample") as sfw:
+        # should not be possible, but who am I?
+        assert sfw.get_datatype() == EnergyDType.ivsample
+    with Reader(h5_path) as sfr:
+        assert sfr.get_datatype() == EnergyDType.ivsample
+
+
+def test_writer_not_modify_datatype(h5_path: Path) -> None:
+    generate_shp_file(h5_path, datatype="ivcurve", window_samples=1024)
+    with Reader(h5_path) as sfr:
+        assert sfr.get_datatype() == EnergyDType.ivcurve
+    with Writer(h5_path, modify_existing=True) as sfw:
+        assert sfw.get_datatype() == EnergyDType.ivcurve
+    with Reader(h5_path) as sfr:
+        assert sfr.get_datatype() == EnergyDType.ivcurve
+
+
+def test_writer_modify_window_samples(h5_path: Path) -> None:
+    generate_shp_file(h5_path, window_samples=1024)
+    with Reader(h5_path) as sfr:
+        assert sfr.get_window_samples() == 1024
+    with Writer(h5_path, modify_existing=True, window_samples=2048) as sfw:
+        # should not be possible, but who am I?
+        assert sfw.get_window_samples() == 2048
+    with Reader(h5_path) as sfr:
+        assert sfr.get_window_samples() == 2048
+
+
+def test_writer_not_modify_window_samples(h5_path: Path) -> None:
+    generate_shp_file(h5_path, window_samples=1024)
+    with Reader(h5_path) as sfr:
+        assert sfr.get_window_samples() == 1024
+    with Writer(h5_path, modify_existing=True) as sfw:
+        assert sfw.get_window_samples() == 1024
+    with Reader(h5_path) as sfr:
+        assert sfr.get_window_samples() == 1024
+
+
+def test_writer_modify_cal_data(h5_path: Path) -> None:
+    cal1 = CalSeries.from_cal(CalHrv(adc_V_Sense=CalibrationPair(gain=1.23)))
+    cal2 = CalSeries.from_cal(CalHrv(adc_V_Sense=CalibrationPair(gain=2.34)))
+    generate_shp_file(h5_path, cal_data=cal1)
+    with Reader(h5_path) as sfr:
+        assert sfr.get_calibration_data().voltage.gain == 1.23
+    with Writer(h5_path, modify_existing=True, cal_data=cal2) as sfw:
+        assert sfw.get_calibration_data().voltage.gain == 2.34
+    with Reader(h5_path) as sfr:
+        assert sfr.get_calibration_data().voltage.gain == 2.34
+
+
+def test_writer_not_modify_cal_data(h5_path: Path) -> None:
+    cal1 = CalSeries.from_cal(CalHrv(adc_V_Sense=CalibrationPair(gain=1.23)))
+    generate_shp_file(h5_path, cal_data=cal1)
+    with Reader(h5_path) as sfr:
+        assert sfr.get_calibration_data().voltage.gain == 1.23
+    with Writer(h5_path, modify_existing=True) as sfw:
+        assert sfw.get_calibration_data().voltage.gain == 1.23
+    with Reader(h5_path) as sfr:
+        assert sfr.get_calibration_data().voltage.gain == 1.23
+
+
+def test_writer_modify_hostname(h5_path: Path) -> None:
+    generate_shp_file(h5_path, hostname="bubu")
+    with Reader(h5_path) as sfr:
+        assert sfr.get_hostname() == "bubu"
+    with Writer(h5_path, modify_existing=True) as sfw:
+        sfw.store_hostname("sheep")
+        assert sfw.get_hostname() == "sheep"
+    with Reader(h5_path) as sfr:
+        assert sfr.get_hostname() == "sheep"
+
+
+def test_writer_not_modify_hostname(h5_path: Path) -> None:
+    generate_shp_file(h5_path, hostname="bubu")
+    with Reader(h5_path) as sfr:
+        assert sfr.get_hostname() == "bubu"
+    with Writer(h5_path, modify_existing=True) as sfw:
+        assert sfw.get_hostname() == "bubu"
+    with Reader(h5_path) as sfr:
+        assert sfr.get_hostname() == "bubu"
+
+
+# TODO: get_config
 
 # TODO:
 #  - test writing different and confirming them