+++ shepherd_data/tests/test_cli_extract.py
@@ -11,30 +11,37 @@ def test_cli_extract_file_full(data_h5: Path) -> None:
         [
             "--verbose",
             "extract",
+            "--start",
+            "1",
+            "--end",
+            "5",
             "--ds-factor",
             "100",
             "--separator",
             ",",
+            "--raw",
             str(data_h5),
         ],
     )
     assert res.exit_code == 0
-    assert data_h5.with_suffix(".downsampled_x100.h5").exists()
-    assert data_h5.with_suffix(".downsampled_x100.data.csv").exists()
+    assert data_h5.with_suffix(".cut_1s000_to_5s000.downsample_x100.h5").exists()
+    assert data_h5.with_suffix(".cut_1s000_to_5s000.downsample_x100.data.csv").exists()
 
 
 def test_cli_extract_file_short(data_h5: Path) -> None:
-    res = CliRunner().invoke(cli, ["-v", "extract", "-f", "200", "-s", ";", str(data_h5)])
+    res = CliRunner().invoke(
+        cli, ["-v", "extract", "-f", "200", "-s", "2", "-e", "4", "--separator", ";", str(data_h5)]
+    )
     assert res.exit_code == 0
-    assert data_h5.with_suffix(".downsampled_x200.h5").exists()
-    assert data_h5.with_suffix(".downsampled_x200.data.csv").exists()
+    assert data_h5.with_suffix(".cut_2s000_to_4s000.downsample_x200.h5").exists()
+    assert data_h5.with_suffix(".cut_2s000_to_4s000.downsample_x200.data.csv").exists()
 
 
 def test_cli_extract_file_min(data_h5: Path) -> None:
     res = CliRunner().invoke(cli, ["-v", "extract", str(data_h5)])
     assert res.exit_code == 0
-    assert data_h5.with_suffix(".downsampled_x1000.h5").exists()
-    assert data_h5.with_suffix(".downsampled_x1000.data.csv").exists()
+    assert data_h5.with_suffix(".downsample_x1000.h5").exists()
+    assert data_h5.with_suffix(".downsample_x1000.data.csv").exists()
 
 
 def test_cli_extract_dir_full(data_h5: Path) -> None:
@@ -46,35 +53,12 @@ def test_cli_extract_dir_full(data_h5: Path) -> None:
             "--verbose",
             "extract",
             "--ds-factor",
-            "2000",
+            "300",
             "--separator",
             ";",
             str(data_h5.parent),
         ],
     )
     assert res.exit_code == 0
-    assert data_h5.with_suffix(".downsampled_x2000.h5").exists()
-    assert data_h5.with_suffix(".downsampled_x2000.data.csv").exists()
-
-
-def test_cli_extract_meta_file_full(data_h5: Path) -> None:
-    res = CliRunner().invoke(cli, ["--verbose", "extract-meta", "--separator", ";", str(data_h5)])
-    assert res.exit_code == 0
-    # TODO: nothing to grab here, add in base-file, same for tests below
-
-
-def test_cli_extract_meta_file_short(data_h5: Path) -> None:
-    res = CliRunner().invoke(cli, ["-v", "extract-meta", "-s", "-", str(data_h5)])
-    assert res.exit_code == 0
-
-
-def test_cli_extract_meta_file_min(data_h5: Path) -> None:
-    res = CliRunner().invoke(cli, ["-v", "extract-meta", "-s", "-", str(data_h5)])
-    assert res.exit_code == 0
-
-
-def test_cli_extract_meta_dir_full(data_h5: Path) -> None:
-    res = CliRunner().invoke(
-        cli, ["--verbose", "extract-meta", "--separator", ";", str(data_h5.parent)]
-    )
-    assert res.exit_code == 0
+    assert data_h5.with_suffix(".downsample_x300.h5").exists()
+    assert data_h5.with_suffix(".downsample_x300.data.csv").exists()